final-lfarley1-jyang4-mle1-main/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
banks==2.4.1
bbox_visualizer==0.3.0
c2==0.1.7
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
dataclasses-json==0.6.7
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
einops==0.8.2
exceptiongroup==1.3.1
fastapi==0.135.2
filelock==3.20.3
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
griffe==1.15.0
griffecli==2.0.0
griffelib==2.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.18
llama-index-cli==0.5.6
llama-index-core==0.14.18
llama-index-embeddings-openai==0.6.0
llama-index-indices-managed-llama-cloud==0.10.0
llama-index-instrumentation==0.5.0
llama-index-llms-openai==0.7.2
llama-index-readers-llama-parse==0.6.0
llama-index-workflows==2.16.1
llama-parse==0.6.54
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.3
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai>=2.16.0
openai-agents==0.12.5
opencv-python==4.13.0.90
packaging==26.0
pillow==12.1.0
platformdirs==4.9.4
postgrest==2.28.3
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
realtime==2.28.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.48
sse-starlette==3.3.3
starlette==0.52.1
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
timm==1.0.24
tinytag==2.2.1
tokenizers==0.21.4
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.2
transformers==4.49.0
triton==3.6.0
typer==0.4.2
typer-slim==0.21.1
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==15.0.1
wheel==0.46.3
wrapt==1.17.3
yarl==1.23.0
zstandard==0.25.0
